TelegramBot/requirements.txt
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
future==0.18.2
idna==2.10
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.0
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2020.12.4
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
